Refactoring long class names.
