+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/domestic/scheduled/payments/api/v3_1_8/GetDomesticScheduledPaymentDetails.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.payment.domestic.payments.api.v3_1_8
+package com.forgerock.uk.openbanking.tests.functional.payment.domestic.scheduled.payments.api.v3_1_8
 
 import assertk.assertThat
 import assertk.assertions.isNotNull
@@ -8,25 +8,25 @@ import com.forgerock.uk.openbanking.support.discovery.getPaymentsApiLinks
 import com.forgerock.uk.openbanking.support.payment.PaymentFactory
 import com.forgerock.uk.openbanking.support.payment.defaultPaymentScopesForAccessToken
 import uk.org.openbanking.datamodel.payment.OBWritePaymentDetailsResponse1
-import uk.org.openbanking.testsupport.payment.OBWriteDomesticConsentTestDataFactory
+import uk.org.openbanking.testsupport.payment.OBWriteDomesticScheduledConsentTestDataFactory
 
-class GetDomesticPaymentDomesticPaymentIdPaymentDetails(
+class GetDomesticScheduledPaymentDetails(
     val version: OBVersion,
     val tppResource: CreateTppCallback.TppResource
 ) {
-    private val createDomesticPaymentApi = CreateDomesticPayment(version, tppResource)
+    private val createDomesticScheduledPayments = CreateDomesticScheduledPayment(version, tppResource)
     private val paymentLinks = getPaymentsApiLinks(version)
     private val paymentApiClient = tppResource.tpp.paymentApiClient
 
-    fun getDomesticPaymentDomesticPaymentIdPaymentDetailsTest() {
+    fun getDomesticScheduledPaymentDomesticPaymentIdPaymentDetailsTest() {
         // Given
-        val consentRequest = OBWriteDomesticConsentTestDataFactory.aValidOBWriteDomesticConsent4()
-        val paymentResponse = createDomesticPaymentApi.submitPayment(consentRequest)
+        val consentRequest = OBWriteDomesticScheduledConsentTestDataFactory.aValidOBWriteDomesticScheduledConsent4()
+        val paymentResponse = createDomesticScheduledPayments.submitPayment(consentRequest)
 
         // When
-        val getDomesticPaymentDetailsUrl = PaymentFactory.urlWithDomesticPaymentId(
-            paymentLinks.GetDomesticPaymentDomesticPaymentIdPaymentDetails,
-            paymentResponse.data.domesticPaymentId
+        val getDomesticPaymentDetailsUrl = PaymentFactory.urlWithDomesticScheduledPaymentId(
+            paymentLinks.GetDomesticScheduledPaymentDomesticPaymentIdPaymentDetails,
+            paymentResponse.data.domesticScheduledPaymentId
         )
         val paymentDetailsResponse = paymentApiClient.sendGetRequest<OBWritePaymentDetailsResponse1>(
             getDomesticPaymentDetailsUrl,
