+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/domestic/scheduled/payments/junit/v3_1_10/GetDomesticScheduledPaymentDetailsTest.kt
@@ -1,23 +1,23 @@
-package com.forgerock.uk.openbanking.tests.functional.payment.domestic.scheduled.payments.junit.v3_1_9
+package com.forgerock.uk.openbanking.tests.functional.payment.domestic.scheduled.payments.junit.v3_1_10
 
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
 import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
-import com.forgerock.uk.openbanking.tests.functional.payment.domestic.scheduled.payments.api.v3_1_8.GetDomesticScheduledPaymentDomesticPaymentIdPaymentDetails
+import com.forgerock.uk.openbanking.tests.functional.payment.domestic.scheduled.payments.api.v3_1_8.GetDomesticScheduledPaymentDetails
 import org.junit.jupiter.api.Test
 
-class GetDomesticScheduledPaymentDomesticPaymentIdPaymentDetailsTest(val tppResource: CreateTppCallback.TppResource) {
+class GetDomesticScheduledPaymentDetailsTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "payments",
-        apiVersion = "v3.1.9",
+        apiVersion = "v3.1.10",
         operations = ["GetDomesticScheduledPayment", "CreateDomesticScheduledPayment", "CreateDomesticScheduledPaymentConsent", "GetDomesticScheduledPaymentConsent", "GetDomesticScheduledPaymentDomesticPaymentIdPaymentDetails"],
         apis = ["domestic-scheduled-payments", "domestic-scheduled-payment-consents"]
     )
     @Test
-    fun getDomesticScheduledPaymentDomesticPaymentIdPaymentDetails_v3_1_9() {
-        GetDomesticScheduledPaymentDomesticPaymentIdPaymentDetails(
-            OBVersion.v3_1_9,
+    fun getDomesticScheduledPaymentDomesticPaymentIdPaymentDetails_v3_1_10() {
+        GetDomesticScheduledPaymentDetails(
+            OBVersion.v3_1_10,
             tppResource
         ).getDomesticScheduledPaymentDomesticPaymentIdPaymentDetailsTest()
     }
