+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/domestic/scheduled/payments/legacy/LegacyGetDomesticScheduledPaymentDetailsTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.payment.domestic.payments.legacy
+package com.forgerock.uk.openbanking.tests.functional.payment.domestic.scheduled.payments.legacy
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
@@ -15,27 +15,25 @@ import com.forgerock.uk.openbanking.support.discovery.payment3_1_3
 import com.forgerock.uk.openbanking.support.discovery.payment3_1_4
 import com.forgerock.uk.openbanking.support.payment.PaymentAS
 import com.forgerock.uk.openbanking.support.payment.PaymentFactory
-import com.forgerock.uk.openbanking.support.payment.PaymentFactory.Companion.copyOBWriteDomestic2DataInitiation
-import com.forgerock.uk.openbanking.support.payment.PaymentFactory.Companion.urlWithDomesticPaymentId
+import com.forgerock.uk.openbanking.support.payment.PaymentFactory.Companion.urlWithDomesticScheduledPaymentId
 import com.forgerock.uk.openbanking.support.payment.PaymentRS
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.payment.*
-import uk.org.openbanking.testsupport.payment.OBWriteDomesticConsentTestDataFactory.aValidOBWriteDomesticConsent3
-import uk.org.openbanking.testsupport.payment.OBWriteDomesticConsentTestDataFactory.aValidOBWriteDomesticConsent4
+import uk.org.openbanking.testsupport.payment.OBWriteDomesticScheduledConsentTestDataFactory.aValidOBWriteDomesticScheduledConsent3
+import uk.org.openbanking.testsupport.payment.OBWriteDomesticScheduledConsentTestDataFactory.aValidOBWriteDomesticScheduledConsent4
 
-class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResource: CreateTppCallback.TppResource) {
-
+class LegacyGetDomesticScheduledPaymentDetailsTest(val tppResource: CreateTppCallback.TppResource) {
     @EnabledIfVersion(
         type = "payments",
         apiVersion = "v3.1.4",
-        operations = ["GetDomesticPayment", "CreateDomesticPayment", "CreateDomesticPaymentConsent", "GetDomesticPaymentConsent", "GetDomesticPaymentDomesticPaymentIdPaymentDetails"],
-        apis = ["domestic-payments", "domestic-payment-consents"]
+        operations = ["GetDomesticScheduledPayment", "CreateDomesticScheduledPayment", "CreateDomesticScheduledPaymentConsent", "GetDomesticScheduledPaymentConsent", "GetDomesticScheduledPaymentDomesticPaymentIdPaymentDetails"],
+        apis = ["domestic-scheduled-payments", "domestic-scheduled-payment-consents"]
     )
     @Test
-    fun getDomesticPaymentDomesticPaymentIdPaymentDetails_v3_1_4() {
+    fun getDomesticScheduledPaymentDomesticPaymentIdPaymentDetails_v3_1_4() {
         // Given
-        val consentRequest = aValidOBWriteDomesticConsent4()
+        val consentRequest = aValidOBWriteDomesticScheduledConsent4()
 
         val signedPayloadConsent =
             signPayloadSubmitPayment(
@@ -44,8 +42,8 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
                 tppResource.tpp.signingKid
             )
 
-        val consent = PaymentRS().consent<OBWriteDomesticConsentResponse4>(
-            payment3_1_4.Links.links.CreateDomesticPaymentConsent,
+        val consent = PaymentRS().consent<OBWriteDomesticScheduledConsentResponse4>(
+            payment3_1_4.Links.links.CreateDomesticScheduledPaymentConsent,
             consentRequest,
             tppResource.tpp,
             OBVersion.v3_1_4,
@@ -68,9 +66,9 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
         // accessToken to get the payment use the grant type client_credentials
         val accessTokenClientCredentials = PaymentRS().getClientCredentialsAccessToken(tppResource.tpp)
 
-        val patchedConsent = PaymentRS().getConsent<OBWriteDomesticConsentResponse4>(
+        val patchedConsent = PaymentRS().getConsent<OBWriteDomesticScheduledConsentResponse4>(
             PaymentFactory.urlWithConsentId(
-                payment3_1_4.Links.links.GetDomesticPaymentConsent,
+                payment3_1_4.Links.links.GetDomesticScheduledPaymentConsent,
                 consent.data.consentId
             ),
             tppResource.tpp,
@@ -83,10 +81,10 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
         assertThat(patchedConsent.data.consentId).isNotEmpty()
         Assertions.assertThat(patchedConsent.data.status.toString()).`is`(Status.consentCondition)
 
-        val paymentSubmissionRequest = OBWriteDomestic2().data(
-            OBWriteDomestic2Data()
+        val paymentSubmissionRequest = OBWriteDomesticScheduled2().data(
+            OBWriteDomesticScheduled2Data()
                 .consentId(patchedConsent.data.consentId)
-                .initiation(copyOBWriteDomestic2DataInitiation(patchedConsent.data.initiation))
+                .initiation(patchedConsent.data.initiation)
         ).risk(patchedConsent.risk)
 
         val signedPayload = signPayloadSubmitPayment(
@@ -95,8 +93,8 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
             tppResource.tpp.signingKid
         )
 
-        val submissionResponse = PaymentRS().submitPayment<OBWriteDomesticResponse4>(
-            payment3_1_4.Links.links.CreateDomesticPayment,
+        val submissionResponse = PaymentRS().submitPayment<OBWriteDomesticScheduledResponse4>(
+            payment3_1_4.Links.links.CreateDomesticScheduledPayment,
             paymentSubmissionRequest,
             accessTokenAuthorizationCode,
             signedPayload,
@@ -107,13 +105,13 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
         assertThat(submissionResponse).isNotNull()
         assertThat(submissionResponse.data).isNotNull()
         assertThat(submissionResponse.data.consentId).isEqualTo(patchedConsent.data.consentId)
-        assertThat(submissionResponse.data.domesticPaymentId).isNotEmpty()
+        assertThat(submissionResponse.data.domesticScheduledPaymentId).isNotEmpty()
 
         // When
         val paymentResult = PaymentRS().getPayment<OBWritePaymentDetailsResponse1>(
-            urlWithDomesticPaymentId(
-                payment3_1_4.Links.links.GetDomesticPaymentDomesticPaymentIdPaymentDetails,
-                submissionResponse.data.domesticPaymentId
+            urlWithDomesticScheduledPaymentId(
+                payment3_1_4.Links.links.GetDomesticScheduledPaymentDomesticPaymentIdPaymentDetails,
+                submissionResponse.data.domesticScheduledPaymentId
             ),
             accessTokenClientCredentials,
             tppResource.tpp
@@ -128,13 +126,13 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
     @EnabledIfVersion(
         type = "payments",
         apiVersion = "v3.1.3",
-        operations = ["GetDomesticPayment", "CreateDomesticPayment", "CreateDomesticPaymentConsent", "GetDomesticPaymentConsent", "GetDomesticPaymentDomesticPaymentIdPaymentDetails"],
-        apis = ["domestic-payments", "domestic-payment-consents"]
+        operations = ["GetDomesticScheduledPayment", "CreateDomesticScheduledPayment", "CreateDomesticScheduledPaymentConsent", "GetDomesticScheduledPaymentConsent", "GetDomesticScheduledPaymentDomesticPaymentIdPaymentDetails"],
+        apis = ["domestic-scheduled-payments", "domestic-scheduled-payment-consents"]
     )
     @Test
-    fun getDomesticPaymentDomesticPaymentIdPaymentDetails_v3_1_3() {
+    fun getDomesticScheduledPaymentDomesticPaymentIdPaymentDetails_v3_1_3() {
         // Given
-        val consentRequest = aValidOBWriteDomesticConsent3()
+        val consentRequest = aValidOBWriteDomesticScheduledConsent3()
 
         val signedPayloadConsent =
             signPayloadSubmitPayment(
@@ -143,8 +141,8 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
                 tppResource.tpp.signingKid
             )
 
-        val consent = PaymentRS().consent<OBWriteDomesticConsentResponse3>(
-            payment3_1_3.Links.links.CreateDomesticPaymentConsent,
+        val consent = PaymentRS().consent<OBWriteDomesticScheduledConsentResponse3>(
+            payment3_1_3.Links.links.CreateDomesticScheduledPaymentConsent,
             consentRequest,
             tppResource.tpp,
             OBVersion.v3_1_3,
@@ -167,9 +165,9 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
         // accessToken to get the payment use the grant type client_credentials
         val accessTokenClientCredentials = PaymentRS().getClientCredentialsAccessToken(tppResource.tpp)
 
-        val patchedConsent = PaymentRS().getConsent<OBWriteDomesticConsentResponse3>(
+        val patchedConsent = PaymentRS().getConsent<OBWriteDomesticScheduledConsentResponse3>(
             PaymentFactory.urlWithConsentId(
-                payment3_1_3.Links.links.GetDomesticPaymentConsent,
+                payment3_1_3.Links.links.GetDomesticScheduledPaymentConsent,
                 consent.data.consentId
             ),
             tppResource.tpp,
@@ -182,10 +180,10 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
         assertThat(patchedConsent.data.consentId).isNotEmpty()
         Assertions.assertThat(patchedConsent.data.status.toString()).`is`(Status.consentCondition)
 
-        val paymentSubmissionRequest = OBWriteDomestic2().data(
-            OBWriteDomestic2Data()
+        val paymentSubmissionRequest = OBWriteDomesticScheduled2().data(
+            OBWriteDomesticScheduled2Data()
                 .consentId(patchedConsent.data.consentId)
-                .initiation(copyOBWriteDomestic2DataInitiation(patchedConsent.data.initiation))
+                .initiation(patchedConsent.data.initiation)
         ).risk(patchedConsent.risk)
 
         val signedPayload =
@@ -196,8 +194,8 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
                 true
             )
 
-        val submissionResponse = PaymentRS().submitPayment<OBWriteDomesticResponse3>(
-            payment3_1_3.Links.links.CreateDomesticPayment,
+        val submissionResponse = PaymentRS().submitPayment<OBWriteDomesticScheduledResponse3>(
+            payment3_1_3.Links.links.CreateDomesticScheduledPayment,
             paymentSubmissionRequest,
             accessTokenAuthorizationCode,
             signedPayload,
@@ -208,13 +206,13 @@ class LegacyGetDomesticPaymentDomesticPaymentIdPaymentDetailsTest(val tppResourc
         assertThat(submissionResponse).isNotNull()
         assertThat(submissionResponse.data).isNotNull()
         assertThat(submissionResponse.data.consentId).isEqualTo(patchedConsent.data.consentId)
-        assertThat(submissionResponse.data.domesticPaymentId).isNotEmpty()
+        assertThat(submissionResponse.data.domesticScheduledPaymentId).isNotEmpty()
 
         // When
         val paymentResult = PaymentRS().getPayment<OBWritePaymentDetailsResponse1>(
-            urlWithDomesticPaymentId(
-                payment3_1_3.Links.links.GetDomesticPaymentDomesticPaymentIdPaymentDetails,
-                submissionResponse.data.domesticPaymentId
+            urlWithDomesticScheduledPaymentId(
+                payment3_1_3.Links.links.GetDomesticScheduledPaymentDomesticPaymentIdPaymentDetails,
+                submissionResponse.data.domesticScheduledPaymentId
             ),
             accessTokenClientCredentials,
             tppResource.tpp
