+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/international/payments/api/v3_1_8/GetInternationalPaymentDetails.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.payment.international.scheduled.payments.api.v3_1_8
+package com.forgerock.uk.openbanking.tests.functional.payment.international.payments.api.v3_1_8
 
 import assertk.assertThat
 import assertk.assertions.isNotNull
@@ -8,26 +8,23 @@ import com.forgerock.uk.openbanking.support.discovery.getPaymentsApiLinks
 import com.forgerock.uk.openbanking.support.payment.PaymentFactory
 import com.forgerock.uk.openbanking.support.payment.defaultPaymentScopesForAccessToken
 import uk.org.openbanking.datamodel.payment.OBExchangeRateType2Code
-import uk.org.openbanking.datamodel.payment.OBWriteInternationalScheduledResponse5
+import uk.org.openbanking.datamodel.payment.OBWriteInternationalResponse5
 import uk.org.openbanking.datamodel.payment.OBWritePaymentDetailsResponse1
-import uk.org.openbanking.testsupport.payment.OBWriteInternationalScheduledConsentTestDataFactory
+import uk.org.openbanking.testsupport.payment.OBWriteInternationalConsentTestDataFactory
 
-class GetInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDetails(
+class GetInternationalPaymentDetails(
     val version: OBVersion,
     val tppResource: CreateTppCallback.TppResource
 ) {
-    private val createInternationalScheduledPayment =
-        CreateInternationalScheduledPayment(version, tppResource)
+    private val createInternationalPayment = CreateInternationalPayment(version, tppResource)
     private val paymentLinks = getPaymentsApiLinks(version)
     private val paymentApiClient = tppResource.tpp.paymentApiClient
-    
-    fun getInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDetails_rateType_AGREED_Test() {
-        // Given
-        val consentRequest =
-            OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5()
-        consentRequest.data.initiation.exchangeRateInformation.rateType = OBExchangeRateType2Code.AGREED
 
-        val paymentResponse = createInternationalScheduledPayment.submitPayment(consentRequest)
+    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_AGREED_Test() {
+        // Given
+        val consentRequest = OBWriteInternationalConsentTestDataFactory.aValidOBWriteInternationalConsent5()
+        consentRequest.data.initiation.exchangeRateInformation.rateType = OBExchangeRateType2Code.AGREED
+        val paymentResponse = createInternationalPayment.submitPayment(consentRequest)
 
         // When
         val result = getPaymentDetails(paymentResponse)
@@ -38,18 +35,18 @@ class GetInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDeta
         assertThat(result.data.paymentStatus).isNotNull()
     }
 
-    fun getInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDetails_rateType_ACTUAL_Test() {
+    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_ACTUAL_Test() {
         // Given
-        val consentRequest =
-            OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5()
+        val consentRequest = OBWriteInternationalConsentTestDataFactory.aValidOBWriteInternationalConsent5()
         consentRequest.data.initiation.exchangeRateInformation.rateType = OBExchangeRateType2Code.ACTUAL
         consentRequest.data.initiation.exchangeRateInformation.exchangeRate = null
         consentRequest.data.initiation.exchangeRateInformation.contractIdentification = null
 
-        val paymentResponse = createInternationalScheduledPayment.submitPayment(consentRequest)
+        val paymentResponse = createInternationalPayment.submitPayment(consentRequest)
 
         // When
         val result = getPaymentDetails(paymentResponse)
+
 
         // Then
         assertThat(result).isNotNull()
@@ -57,18 +54,18 @@ class GetInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDeta
         assertThat(result.data.paymentStatus).isNotNull()
     }
 
-    fun getInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDetails_rateType_INDICATIVE_Test() {
+    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_INDICATIVE_Test() {
         // Given
-        val consentRequest =
-            OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5()
+        val consentRequest = OBWriteInternationalConsentTestDataFactory.aValidOBWriteInternationalConsent5()
         consentRequest.data.initiation.exchangeRateInformation.rateType = OBExchangeRateType2Code.INDICATIVE
         consentRequest.data.initiation.exchangeRateInformation.exchangeRate = null
         consentRequest.data.initiation.exchangeRateInformation.contractIdentification = null
 
-        val paymentResponse = createInternationalScheduledPayment.submitPayment(consentRequest)
+        val paymentResponse = createInternationalPayment.submitPayment(consentRequest)
 
         // When
         val result = getPaymentDetails(paymentResponse)
+
 
         // Then
         assertThat(result).isNotNull()
@@ -76,15 +73,15 @@ class GetInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDeta
         assertThat(result.data.paymentStatus).isNotNull()
     }
 
-    fun getInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDetails_Test() {
+    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_mandatoryFields_Test() {
         // Given
         val consentRequest =
-            OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5MandatoryFields()
-
-        val paymentResponse = createInternationalScheduledPayment.submitPayment(consentRequest)
+            OBWriteInternationalConsentTestDataFactory.aValidOBWriteInternationalConsent5MandatoryFields()
+        val paymentResponse = createInternationalPayment.submitPayment(consentRequest)
 
         // When
         val result = getPaymentDetails(paymentResponse)
+
 
         // Then
         assertThat(result).isNotNull()
@@ -92,10 +89,10 @@ class GetInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDeta
         assertThat(result.data.paymentStatus).isNotNull()
     }
 
-    private fun getPaymentDetails(paymentResponse: OBWriteInternationalScheduledResponse5): OBWritePaymentDetailsResponse1 {
-        val getInternationalPaymentDetails = PaymentFactory.urlWithInternationalScheduledPaymentId(
-            paymentLinks.GetInternationalScheduledPaymentPaymentIdPaymentDetails,
-            paymentResponse.data.internationalScheduledPaymentId
+    private fun getPaymentDetails(paymentResponse: OBWriteInternationalResponse5): OBWritePaymentDetailsResponse1 {
+        val getInternationalPaymentDetails = PaymentFactory.urlWithInternationalPaymentId(
+            paymentLinks.GetInternationalPaymentInternationalPaymentIdPaymentDetails,
+            paymentResponse.data.internationalPaymentId
         )
         return paymentApiClient.sendGetRequest(
             getInternationalPaymentDetails,
