+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/international/payments/junit/v3_1_8/GetInternationalPaymentDetailsTest.kt
@@ -1,63 +1,67 @@
-package com.forgerock.uk.openbanking.tests.functional.payment.international.payments.junit.v3_1_9
+package com.forgerock.uk.openbanking.tests.functional.payment.international.payments.junit.v3_1_8
 
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
 import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
-import com.forgerock.uk.openbanking.tests.functional.payment.international.payments.api.v3_1_8.GetInternationalPaymentInternationalPaymentIdPaymentDetails
+import com.forgerock.uk.openbanking.tests.functional.payment.international.payments.api.v3_1_8.GetInternationalPaymentDetails
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class GetInternationalPaymentInternationalPaymentIdPaymentDetailsTest(val tppResource: CreateTppCallback.TppResource) {
+class GetInternationalPaymentDetailsTest(val tppResource: CreateTppCallback.TppResource) {
 
-    lateinit var getInternationalPaymentInternationalPaymentIdPaymentDetails: GetInternationalPaymentInternationalPaymentIdPaymentDetails
+    lateinit var getInternationalPaymentInternationalPaymentIdPaymentDetails: GetInternationalPaymentDetails
 
     @BeforeEach
     fun setUp() {
         getInternationalPaymentInternationalPaymentIdPaymentDetails =
-            GetInternationalPaymentInternationalPaymentIdPaymentDetails(OBVersion.v3_1_9, tppResource)
+            GetInternationalPaymentDetails(OBVersion.v3_1_8, tppResource)
     }
 
     @EnabledIfVersion(
         type = "payments",
-        apiVersion = "v3.1.9",
+        apiVersion = "v3.1.8",
         operations = ["GetInternationalPayment", "CreateInternationalPayment", "CreateInternationalPaymentConsent", "GetInternationalPaymentConsent", "GetInternationalPaymentInternationalPaymentIdPaymentDetails"],
-        apis = ["international-payments", "international-payment-consents"]
+        apis = ["international-payments", "international-payment-consents"],
+        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
     )
     @Test
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_AGREED_v3_1_9() {
+    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_AGREED_v3_1_8() {
         getInternationalPaymentInternationalPaymentIdPaymentDetails.getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_AGREED_Test()
     }
 
     @EnabledIfVersion(
         type = "payments",
-        apiVersion = "v3.1.9",
+        apiVersion = "v3.1.8",
         operations = ["GetInternationalPayment", "CreateInternationalPayment", "CreateInternationalPaymentConsent", "GetInternationalPaymentConsent", "GetInternationalPaymentInternationalPaymentIdPaymentDetails"],
-        apis = ["international-payments", "international-payment-consents"]
+        apis = ["international-payments", "international-payment-consents"],
+        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
     )
     @Test
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_ACTUAL_v3_1_9() {
+    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_ACTUAL_v3_1_8() {
         getInternationalPaymentInternationalPaymentIdPaymentDetails.getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_ACTUAL_Test()
     }
 
     @EnabledIfVersion(
         type = "payments",
-        apiVersion = "v3.1.9",
+        apiVersion = "v3.1.8",
         operations = ["GetInternationalPayment", "CreateInternationalPayment", "CreateInternationalPaymentConsent", "GetInternationalPaymentConsent", "GetInternationalPaymentInternationalPaymentIdPaymentDetails"],
-        apis = ["international-payments", "international-payment-consents"]
+        apis = ["international-payments", "international-payment-consents"],
+        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
     )
     @Test
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_INDICATIVE_v3_1_9() {
+    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_INDICATIVE_v3_1_8() {
         getInternationalPaymentInternationalPaymentIdPaymentDetails.getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_INDICATIVE_Test()
     }
 
     @EnabledIfVersion(
         type = "payments",
-        apiVersion = "v3.1.9",
+        apiVersion = "v3.1.8",
         operations = ["GetInternationalPayment", "CreateInternationalPayment", "CreateInternationalPaymentConsent", "GetInternationalPaymentConsent", "GetInternationalPaymentInternationalPaymentIdPaymentDetails"],
-        apis = ["international-payments", "international-payment-consents"]
+        apis = ["international-payments", "international-payment-consents"],
+        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
     )
     @Test
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_mandatoryFields_v3_1_9() {
+    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_mandatoryFields_v3_1_8() {
         getInternationalPaymentInternationalPaymentIdPaymentDetails.getInternationalPaymentInternationalPaymentIdPaymentDetails_mandatoryFields_Test()
     }
 }
