+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/international/scheduled/payments/api/v3_1_8/GetInternationalScheduledPaymentDetails.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.payment.international.payments.api.v3_1_8
+package com.forgerock.uk.openbanking.tests.functional.payment.international.scheduled.payments.api.v3_1_8
 
 import assertk.assertThat
 import assertk.assertions.isNotNull
@@ -8,23 +8,26 @@ import com.forgerock.uk.openbanking.support.discovery.getPaymentsApiLinks
 import com.forgerock.uk.openbanking.support.payment.PaymentFactory
 import com.forgerock.uk.openbanking.support.payment.defaultPaymentScopesForAccessToken
 import uk.org.openbanking.datamodel.payment.OBExchangeRateType2Code
-import uk.org.openbanking.datamodel.payment.OBWriteInternationalResponse5
+import uk.org.openbanking.datamodel.payment.OBWriteInternationalScheduledResponse5
 import uk.org.openbanking.datamodel.payment.OBWritePaymentDetailsResponse1
-import uk.org.openbanking.testsupport.payment.OBWriteInternationalConsentTestDataFactory
+import uk.org.openbanking.testsupport.payment.OBWriteInternationalScheduledConsentTestDataFactory
 
-class GetInternationalPaymentInternationalPaymentIdPaymentDetails(
+class GetInternationalScheduledPaymentDetails(
     val version: OBVersion,
     val tppResource: CreateTppCallback.TppResource
 ) {
-    private val createInternationalPayment = CreateInternationalPayment(version, tppResource)
+    private val createInternationalScheduledPayment =
+        CreateInternationalScheduledPayment(version, tppResource)
     private val paymentLinks = getPaymentsApiLinks(version)
     private val paymentApiClient = tppResource.tpp.paymentApiClient
-
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_AGREED_Test() {
+    
+    fun getInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDetails_rateType_AGREED_Test() {
         // Given
-        val consentRequest = OBWriteInternationalConsentTestDataFactory.aValidOBWriteInternationalConsent5()
+        val consentRequest =
+            OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5()
         consentRequest.data.initiation.exchangeRateInformation.rateType = OBExchangeRateType2Code.AGREED
-        val paymentResponse = createInternationalPayment.submitPayment(consentRequest)
+
+        val paymentResponse = createInternationalScheduledPayment.submitPayment(consentRequest)
 
         // When
         val result = getPaymentDetails(paymentResponse)
@@ -35,18 +38,18 @@ class GetInternationalPaymentInternationalPaymentIdPaymentDetails(
         assertThat(result.data.paymentStatus).isNotNull()
     }
 
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_ACTUAL_Test() {
+    fun getInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDetails_rateType_ACTUAL_Test() {
         // Given
-        val consentRequest = OBWriteInternationalConsentTestDataFactory.aValidOBWriteInternationalConsent5()
+        val consentRequest =
+            OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5()
         consentRequest.data.initiation.exchangeRateInformation.rateType = OBExchangeRateType2Code.ACTUAL
         consentRequest.data.initiation.exchangeRateInformation.exchangeRate = null
         consentRequest.data.initiation.exchangeRateInformation.contractIdentification = null
 
-        val paymentResponse = createInternationalPayment.submitPayment(consentRequest)
+        val paymentResponse = createInternationalScheduledPayment.submitPayment(consentRequest)
 
         // When
         val result = getPaymentDetails(paymentResponse)
-
 
         // Then
         assertThat(result).isNotNull()
@@ -54,18 +57,18 @@ class GetInternationalPaymentInternationalPaymentIdPaymentDetails(
         assertThat(result.data.paymentStatus).isNotNull()
     }
 
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_rateType_INDICATIVE_Test() {
+    fun getInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDetails_rateType_INDICATIVE_Test() {
         // Given
-        val consentRequest = OBWriteInternationalConsentTestDataFactory.aValidOBWriteInternationalConsent5()
+        val consentRequest =
+            OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5()
         consentRequest.data.initiation.exchangeRateInformation.rateType = OBExchangeRateType2Code.INDICATIVE
         consentRequest.data.initiation.exchangeRateInformation.exchangeRate = null
         consentRequest.data.initiation.exchangeRateInformation.contractIdentification = null
 
-        val paymentResponse = createInternationalPayment.submitPayment(consentRequest)
+        val paymentResponse = createInternationalScheduledPayment.submitPayment(consentRequest)
 
         // When
         val result = getPaymentDetails(paymentResponse)
-
 
         // Then
         assertThat(result).isNotNull()
@@ -73,15 +76,15 @@ class GetInternationalPaymentInternationalPaymentIdPaymentDetails(
         assertThat(result.data.paymentStatus).isNotNull()
     }
 
-    fun getInternationalPaymentInternationalPaymentIdPaymentDetails_mandatoryFields_Test() {
+    fun getInternationalScheduledPaymentInternationalScheduledPaymentIdPaymentDetails_Test() {
         // Given
         val consentRequest =
-            OBWriteInternationalConsentTestDataFactory.aValidOBWriteInternationalConsent5MandatoryFields()
-        val paymentResponse = createInternationalPayment.submitPayment(consentRequest)
+            OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5MandatoryFields()
+
+        val paymentResponse = createInternationalScheduledPayment.submitPayment(consentRequest)
 
         // When
         val result = getPaymentDetails(paymentResponse)
-
 
         // Then
         assertThat(result).isNotNull()
@@ -89,10 +92,10 @@ class GetInternationalPaymentInternationalPaymentIdPaymentDetails(
         assertThat(result.data.paymentStatus).isNotNull()
     }
 
-    private fun getPaymentDetails(paymentResponse: OBWriteInternationalResponse5): OBWritePaymentDetailsResponse1 {
-        val getInternationalPaymentDetails = PaymentFactory.urlWithInternationalPaymentId(
-            paymentLinks.GetInternationalPaymentInternationalPaymentIdPaymentDetails,
-            paymentResponse.data.internationalPaymentId
+    private fun getPaymentDetails(paymentResponse: OBWriteInternationalScheduledResponse5): OBWritePaymentDetailsResponse1 {
+        val getInternationalPaymentDetails = PaymentFactory.urlWithInternationalScheduledPaymentId(
+            paymentLinks.GetInternationalScheduledPaymentPaymentIdPaymentDetails,
+            paymentResponse.data.internationalScheduledPaymentId
         )
         return paymentApiClient.sendGetRequest(
             getInternationalPaymentDetails,
