+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/international/scheduled/payments/legacy/LegacyGetInternationalScheduledPaymentDetailsTest.kt
