+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/international/standing/orders/junit/v3_1_9/GetInternationalStandingOrderDetailsTest.kt
@@ -1,43 +1,41 @@
-package com.forgerock.uk.openbanking.tests.functional.payment.international.standing.orders.junit.v3_1_8
+package com.forgerock.uk.openbanking.tests.functional.payment.international.standing.orders.junit.v3_1_9
 
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
 import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
-import com.forgerock.uk.openbanking.tests.functional.payment.international.standing.orders.api.v3_1_8.GetInternationalStandingOrderInternationalStandingOrderIdPaymentDetails
+import com.forgerock.uk.openbanking.tests.functional.payment.international.standing.orders.api.v3_1_8.GetInternationalStandingOrderDetails
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class GetInternationalStandingOrderInternationalStandingOrderIdPaymentDetailsTest(val tppResource: CreateTppCallback.TppResource) {
+class GetInternationalStandingOrderDetailsTest(val tppResource: CreateTppCallback.TppResource) {
 
-    lateinit var getInternationalStandingOrderInternationalStandingOrderIdPaymentDetails: GetInternationalStandingOrderInternationalStandingOrderIdPaymentDetails
+    lateinit var getInternationalStandingOrderInternationalStandingOrderIdPaymentDetails: GetInternationalStandingOrderDetails
 
     @BeforeEach
     fun setUp() {
         getInternationalStandingOrderInternationalStandingOrderIdPaymentDetails =
-            GetInternationalStandingOrderInternationalStandingOrderIdPaymentDetails(OBVersion.v3_1_8, tppResource)
+            GetInternationalStandingOrderDetails(OBVersion.v3_1_9, tppResource)
     }
 
     @EnabledIfVersion(
         type = "payments",
-        apiVersion = "v3.1.8",
+        apiVersion = "v3.1.9",
         operations = ["GetInternationalStandingOrder", "CreateInternationalStandingOrder", "CreateInternationalStandingOrderConsent", "GetInternationalStandingOrderConsent", "GetInternationalStandingOrderInternationalStandingOrderIdPaymentDetails"],
-        apis = ["international-standing-orders", "international-standing-order-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
+        apis = ["international-standing-orders", "international-standing-order-consents"]
     )
     @Test
-    fun getInternationalStandingOrderInternationalStandingOrderIdPaymentDetails_v3_1_8() {
+    fun getInternationalStandingOrderInternationalStandingOrderIdPaymentDetails_v3_1_9() {
         getInternationalStandingOrderInternationalStandingOrderIdPaymentDetails.getInternationalStandingOrderInternationalStandingOrderIdPaymentDetailsTest()
     }
 
     @EnabledIfVersion(
         type = "payments",
-        apiVersion = "v3.1.8",
+        apiVersion = "v3.1.9",
         operations = ["GetInternationalStandingOrder", "CreateInternationalStandingOrder", "CreateInternationalStandingOrderConsent", "GetInternationalStandingOrderConsent", "GetInternationalStandingOrderInternationalStandingOrderIdPaymentDetails"],
-        apis = ["international-standing-orders", "international-standing-order-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
+        apis = ["international-standing-orders", "international-standing-order-consents"]
     )
     @Test
-    fun getInternationalStandingOrderInternationalStandingOrderIdPaymentDetails_mandatoryFields_v3_1_8() {
+    fun getInternationalStandingOrderInternationalStandingOrderIdPaymentDetails_mandatoryFields_v3_1_9() {
         getInternationalStandingOrderInternationalStandingOrderIdPaymentDetails.getInternationalStandingOrderInternationalStandingOrderIdPaymentDetails_mandatoryFieldsTest()
     }
 }
